BACKEND/backend/pay_flow_project/biometric_auth/utils.py
import os
import base64
import numpy as np
from io import BytesIO
from PIL import Image

class BiometricProcessor:
    @staticmethod
    def process_face_image(base64_image):
        """
        Process a base64 encoded image for face recognition.
        In a real implementation, this would use face recognition libraries.
        """
        try:
            # Decode base64 image
            image_data = base64.b64decode(base64_image.split(',')[1])
            image = Image.open(BytesIO(image_data))
            
            # Here you would typically:
            # 1. Detect faces
            # 2. Extract facial features
            # 3. Generate a template for matching
            
            # For this example, just return a placeholder
            return b"face_template_placeholder"
        except Exception as e:
            print(f"Error processing face image: {e}")
            return None
    
    @staticmethod
    def process_fingerprint(base64_fingerprint):
        """
        Process a base64 encoded fingerprint image.
        In a real implementation, this would use fingerprint recognition libraries.
        """
        try:
            # Decode base64 fingerprint
            fingerprint_data = base64.b64decode(base64_fingerprint.split(',')[1])
            
            # Here you would typically:
            # 1. Extract minutiae points
            # 2. Generate a template for matching
            
            # For this example, just return a placeholder
            return b"fingerprint_template_placeholder"
        except Exception as e:
            print(f"Error processing fingerprint: {e}")
            return None
    
    @staticmethod
    def verify_face(stored_template, new_template):
        """
        Compare face templates to verify identity.
        In a real implementation, this would use proper comparison algorithms.
        """
        # This is a placeholder. In a real system, you would:
        # 1. Compare features between templates
        # 2. Calculate similarity score
        # 3. Return match decision based on threshold
        
        # For demo purposes, always return True or implement basic comparison
        return True
    
    @staticmethod
    def verify_fingerprint(stored_template, new_template):
        """
        Compare fingerprint templates to verify identity.
        In a real implementation, this would use proper comparison algorithms.
        """
        # This is a placeholder. In a real system, you would:
        # 1. Compare minutiae points between templates
        # 2. Calculate similarity score
        # 3. Return match decision based on threshold
        
        # For demo purposes, always return True or implement basic comparison
        return True