BACKEND/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
numpy==2.2.3
pillow==11.1.0
psutil==7.0.0
PyJWT==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
urllib3==2.3.0
